pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langgraph-agent"
version = "0.1.0"
description = "LangGraph Agent with Router, Tools, RAG and Memory (DeepSeek + FAISS)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["langgraph", "agent", "llm", "rag", "tools", "deepseek", "faiss", "vector-db"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "langchain>=0.3.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "langchain-community>=0.3.0,<1.0.0",
    "langgraph>=0.2.13,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pyyaml>=6.0,<7.0.0",
    "faiss-cpu>=1.9.0,<2.0.0",
    "sentence-transformers>=3.0.0,<4.0.0",
    "transformers>=4.40.0,<5.0.0",
    "numpy<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.0.0,<6.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.4.0,<1.0.0",
    "mypy>=1.10.0,<2.0.0",
]
gpu = [
    "faiss-gpu>=1.9.0,<2.0.0",
]

[project.scripts]
langgraph-agent = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/langgraph-agent"
Documentation = "https://github.com/yourusername/langgraph-agent#readme"
Repository = "https://github.com/yourusername/langgraph-agent.git"
Issues = "https://github.com/yourusername/langgraph-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/config.yaml",
    "/data",
    "/tests",
    "/scripts",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
target-version = ['py310']
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | env
  | venv
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument (common in langgraph nodes)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
return_types = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    'venv',
    '.venv',
    'env',
    '.env',
    'dist',
    'build',
    'node_modules',
]

[[tool.mypy.overrides]]
module = ["langchain.*", "langgraph.*", "faiss.*", "transformers.*", "sentence_transformers.*"]
ignore_missing_imports = true
